+++ mojo/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.2.35"
+__version__ = "1.2.36"
 
 from mojo.helpers.response import JsonResponse